Split up OptionsArray
